Same base for distant and local IiifLayer
