--- app/components/LocalIiifLayer/tests/index.test.ts
@@ -1,11 +1,0 @@
-// import React from 'react';
-// import { mount } from 'enzyme';
-// import { enzymeFind } from 'styled-components/test-utils';
-
-// import IiifLayer from '../index';
-
-describe('<IiifLayer />', () => {
-  it('Expect to have unit tests specified', () => {
-    expect(true).toEqual(false);
-  });
-});
