+++ babel.config.js
@@ -9,7 +9,6 @@ module.exports = {
     '@babel/preset-react',
   ],
   plugins: [
-    'styled-components',
     '@babel/plugin-proposal-class-properties',
     '@babel/plugin-syntax-dynamic-import',
   ],
