+++ package.json
@@ -136,7 +136,6 @@
     "sanitize.css": "10.0.0",
     "sat": "^0.7.1",
     "save-file": "^2.3.1",
-    "styled-components": "4.1.3",
     "svg-intersections": "^0.4.0",
     "typesafe-actions": "^4.2.1",
     "typings-for-css-modules-loader": "^1.7.0",
@@ -248,7 +247,6 @@
     "tslint-loader": "^3.5.4",
     "tslint-react": "^4.0.0",
     "typescript": "^3.4.5",
-    "typescript-plugin-styled-components": "^1.3.0",
     "url-loader": "1.1.2",
     "webpack": "4.29.6",
     "webpack-cli": "3.2.3",
